refactor: move impl snapshot6 from stream for impl_/stream
